style: 💄 rename Layout.Home and Home component
