+++ src/routes/index.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from "react-router-dom";
 
-import LayoutHome from "../components/organisms/Layout.Home";
+import Home from "./Home";
 import Category from "./Category";
 import SingleEvent from "./SingleEvent";
 import NotFound from "./NotFound";
@@ -8,7 +8,7 @@ import NotFound from "./NotFound";
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <LayoutHome />,
+    element: <Home />,
     errorElement: <NotFound />,
   },
   {
